QA/py/requirements.txt
#
# Note: Tox does not manage dependencies changes. Remove .tox directory after changes here.
#
# START OF BACK-END REQUIREMENTS
# Typings there depend on SQLA 2.0 interface
SQLAlchemy[mypy]==1.4.31
# sqlalchemy2-stubs installed with [mypy] above
sqlalchemy_views==0.3.1
# When SQLAlchemy moves to 2.0 we'll be able to use psycopg3
#psycopg[binary]==3.0.8
psycopg2-binary==2.9.3
# pgvector for similarity search
pgvector==0.2.4
# Image processing during import
Pillow==8.1.0
#Pillow==10.4.0 # Last with python3.8
# astral needs pytz
astral==2.2
pytz==2021.1
types-pytz==2021.1.2
# Needed by TF EXACTLY
numpy~=1.19.2
# Backbone of the app
fastapi==0.73.0
fastapi-utils==0.2.1
# Last (April 2023) pydantic
pydantic[mypy]==1.10.7
# Speed up openapi response serialization
orjson==3.6.6
types-orjson==3.6.2
#graphene==2.1.8
uvloop==0.16.0
httptools==0.2.0
# With 0.16.0: Error "INFO:     ASGI 'lifespan' protocol appears unsupported." and the process
# does not exit. Seen with 0.15.0 as well.
# Note: there was a reg. in 0.17.0 with file upload, not visible during UT. So, to test manually.
uvicorn==0.17.4
gunicorn==20.1.0
httpx==0.28.1
lxml==4.6.3
lxml-stubs==0.2.0
requests==2.26.0
types-requests==2.25.6
python-multipart==0.0.5
# Auth validation
itsdangerous==2.0.1
passlib==1.7.4
# A bit of html
jinja2==3.0.1
aiofiles==0.7.0
# mime type for import
magic-rs==0.2.0
# Analyzis of XLSX data sources
openpyxl==3.0.9
openpyxl-stubs==0.1.21
# Image processing libs.
# Machine learning with no GPU
scikit-learn==1.0
# Machine learning with GPU. NOT automatically regression-tested so to upgrade with caution.
#tensorflow==2.6.0
#tensorflow-hub==0.12.0
#tensorflow_addons==0.14.0
# Used same place as TF, CNN generation
pandas==1.3.3
##lycon==0.2.0 # Conflicts with the version required by TF
#opencv-python-headless==4.5.3.56
#imgaug==0.4.0
# OpenId lib, latest 1.6.6 Requires-Python >=3.9
authlib==1.3.2
# DB upgrade
alembic==1.7.5
# CLI tools, for "cmds" subdirectory
typer[all]==0.4.0
# https://github.com/tiangolo/typer/issues/377
click==8.0.4
# More recent mypy
mypy==1.2.0
# END OF BACK-END REQUIREMENTS
# Testing & QA additions
tox==4.25.0
pytest==8.3.5
pytest-mock==3.14.1
pytest-asyncio==0.24.0
coverage==7.6.1
flake8==3.9.2
radon==5.1.0
deepdiff==5.5.0
# For openapi stubs generation
openapi-generator-cli==4.3.1
# For python API client generator
PyYAML==5.4.1
stringcase==1.2.0
# For python API client generated classes
# Fails to build on GH due to obscure compilation issue. Nevermind for the moment.
#backports-datetime-fromisoformat==1.0.0
